Refactor test dir structure
